move all mixins to one package
delete chesttracker.mixins.json and edit fabric.mod.json to reflect that

move contents of ChestTracker.onInitializeClient to the same method on SimpleShulkerPreviewMod
