--- src/main/resources/chesttracker.mixins.json
@@ -1,16 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "red.jackf.chesttracker_adapted.mixins",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [],
-  "client": [
-    "AccessorMinecraftServer",
-    "MixinClientConnection",
-    "MixinRealmsMainScreen",
-    "MixinShulkerBoxBlock"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
