+++ src/main/resources/simpleshulkerpreview.mixins.json
@@ -5,7 +5,11 @@
   "compatibilityLevel": "JAVA_17",
   "mixins": [],
   "client": [
+    "AccessorMinecraftServer",
     "DrawContextMixin",
+    "MixinClientConnection",
+    "MixinRealmsMainScreen",
+    "MixinShulkerBoxBlock",
     "ShulkerBoxBlockEntityRendererMixin"
   ],
   "injectors": {
